requirements.txt
fastapi
uvicorn
python-telegram-bot
firebase-admin
google-generativeai
pydantic
